[FIX/#102] onboarding xml 네이밍 재설정
