+++ app/src/main/res/layout/fragment_add_friend.xml
@@ -80,7 +80,7 @@
             app:layout_constraintTop_toTopOf="parent" />
 
         <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/rv_freind_list"
+            android:id="@+id/rv_friend_list"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:orientation="vertical"
@@ -88,10 +88,10 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/tv_friend_howmany"
-            tools:listitem="@layout/item_addfriend" />
+            tools:listitem="@layout/item_add_friend" />
 
         <android.widget.Button
-            android:id="@+id/btn_addfriend_next"
+            android:id="@+id/btn_add_friend_next"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:layout_marginHorizontal="16dp"
