+++ buildSrc/src/main/kotlin/Dependencies.kt
@@ -77,8 +77,8 @@ object ThirdPartyDependencies {
     const val kakaoAuth = "com.kakao.sdk:v2-auth:${Versions.kakaoVersion}"
     const val kakaoTalk = "com.kakao.sdk:v2-talk:${Versions.kakaoVersion}"
     const val kakaoShare = "com.kakao.sdk:v2-share:${Versions.kakaoVersion}"
-    // circle scale dot indicator
     const val circleIndicator = "me.relex:circleindicator:${Versions.circleIndicatorVersion}"
+    const val shimmer = "com.facebook.shimmer:shimmer:${Versions.shimmerVersion}"
 }
 
 object ClassPathPlugins {
